refactor: split cmd/tfdrift/main.go by responsibility (fixes #86)
Decompose the 332-line main.go into focused files:
- main.go: CLI entry point and flag parsing
- server.go: HTTP server setup and drift detection dispatch
- import.go: Configuration auto-detection and overrides
- approval.go: Approval workflow subcommands

No functional changes — pure structural refactoring.

Closes #86
